Rename `PublishingEventPipeline` to `DocumentSyncWorker`
This more clearly reflects what it does, now that we have decided to
keep it as part of the same app for now.
