+++ lib/document_sync_worker/configuration.rb
@@ -1,9 +1,9 @@
-module PublishingEventPipeline
+module DocumentSyncWorker
   class Configuration
     attr_accessor :logger, :message_queue_name, :repository
 
     def initialize
-      @logger = Logger.new($stdout, progname: "PublishingEventPipeline")
+      @logger = Logger.new($stdout, progname: "DocumentSyncWorker")
     end
   end
 end
