+++ spec/lib/document_sync_worker/document_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe PublishingEventPipeline::Document do
+RSpec.describe DocumentSyncWorker::Document do
   describe ".for" do
     subject(:document) { described_class.for(document_hash) }
 
@@ -8,14 +8,14 @@ RSpec.describe PublishingEventPipeline::Document do
       context "when the document type is #{document_type}" do
         let(:document_type) { document_type }
 
-        it { is_expected.to be_a(PublishingEventPipeline::Document::Unpublish) }
+        it { is_expected.to be_a(DocumentSyncWorker::Document::Unpublish) }
       end
     end
 
     context "when the document type is not one of the unpublish document types" do
       let(:document_type) { "anything-else" }
 
-      it { is_expected.to be_a(PublishingEventPipeline::Document::Publish) }
+      it { is_expected.to be_a(DocumentSyncWorker::Document::Publish) }
     end
   end
 end
